pyproject.toml
[project]
name = "mcts-self-refine"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Yuya Fujiwara", email = "y.fujiwara@petapa.jp" }
]
dependencies = [
    "pydantic>=2.8.0",
    "openai>=1.35.7",
    "instructor>=1.3.4",
    "manim>=0.18.1",
    "ollama>=0.2.1",
    "python-dotenv>=1.0.1",
    "selenium>=4.22.0",
    "janome>=0.5.0",
]
readme = "README.md"
requires-python = ">= 3.9"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]
